requirements.txt
fastapi==0.110.0
pydantic==2.6.4
uvicorn==0.29.0
gunicorn==21.2.0
httpx==0.27.0
python-multipart==0.0.9
python-dotenv==1.0.1
redisai==1.3.0
numpy==1.23.5
pandas==2.1.0
tensorflow==2.14.0
librosa==0.10.1
resampy==0.4.2
essentia==2.1b6.dev1110
scikit-learn==1.3.0
matplotlib==3.8.0
pretty_midi==0.2.10
imbalanced-learn==0.12.0
